--- examples/o1_web_crawler /requirements.txt
@@ -1,3 +1,0 @@
-firecrawl-py
-python-dotenv
-openai
