+++ examples/o1_web_crawler/requirements.txt
@@ -1,0 +1,3 @@
+firecrawl-py
+python-dotenv
+openai
